+++ R/tidiers_bats.R
@@ -7,8 +7,8 @@
 #' User can supply the original data, which returns the data + augmented columns.
 #' @param rename_index Used with `sw_augment` only.
 #' A string representing the name of the index generated.
-#' @param timekit_idx Used with `sw_augment` and `sw_tidy_decomp`.
-#' When `TRUE`, uses a timekit index (irregular, typically date or datetime) if present.
+#' @param timetk_idx Used with `sw_augment` and `sw_tidy_decomp`.
+#' When `TRUE`, uses a timetk index (irregular, typically date or datetime) if present.
 #' @param ... Additional parameters (not used)
 #'
 #'
@@ -112,13 +112,13 @@ sw_glance.bats <- function(x, ...) {
 #'   * `.resid`: The residual values from the model
 #'
 #' @export
-sw_augment.bats <- function(x, data = NULL, rename_index = "index", timekit_idx = FALSE, ...) {
+sw_augment.bats <- function(x, data = NULL, rename_index = "index", timetk_idx = FALSE, ...) {
 
-    # Check timekit_idx
-    if (timekit_idx) {
-        if (!has_timekit_idx(x)) {
-            warning("Object has no timekit index. Using default index.")
-            timekit_idx = FALSE
+    # Check timetk_idx
+    if (timetk_idx) {
+        if (!has_timetk_idx(x)) {
+            warning("Object has no timetk index. Using default index.")
+            timetk_idx = FALSE
         }
     }
 
@@ -126,14 +126,14 @@ sw_augment.bats <- function(x, data = NULL, rename_index = "index", timekit_idx 
     ret <- tk_tbl(cbind(.actual = x$y, .fitted = x$fitted.values, .resid = x$y - x$fitted.values),
                   rename_index = rename_index, silent = TRUE)
 
-    # Apply timekit index if selected
-    if (timekit_idx) {
-        idx <- tk_index(x, timekit_idx = TRUE)
+    # Apply timetk index if selected
+    if (timetk_idx) {
+        idx <- tk_index(x, timetk_idx = TRUE)
         ret[, rename_index] <- idx
     }
 
     # Augment columns if necessary
-    ret <- sw_augment_columns(ret, data, rename_index, timekit_idx)
+    ret <- sw_augment_columns(ret, data, rename_index, timetk_idx)
 
     return(ret)
 
@@ -151,13 +151,13 @@ sw_augment.bats <- function(x, data = NULL, rename_index = "index", timekit_idx 
 #'   * `season`: The seasonal component (Not always present)
 #'
 #' @export
-sw_tidy_decomp.bats <- function(x, timekit_idx = FALSE, rename_index = "index", ...) {
+sw_tidy_decomp.bats <- function(x, timetk_idx = FALSE, rename_index = "index", ...) {
 
-    # Check timekit_idx
-    if (timekit_idx) {
-        if (!has_timekit_idx(x)) {
-            warning("Object has no timekit index. Using default index.")
-            timekit_idx = FALSE
+    # Check timetk_idx
+    if (timetk_idx) {
+        if (!has_timetk_idx(x)) {
+            warning("Object has no timetk index. Using default index.")
+            timetk_idx = FALSE
         }
     }
 
@@ -173,15 +173,15 @@ sw_tidy_decomp.bats <- function(x, timekit_idx = FALSE, rename_index = "index", 
             dplyr::mutate(seasadj = observed - season)
     }
 
-    # Apply timekit index if selected
-    if (timekit_idx) {
-        idx <- tk_index(x, timekit_idx = TRUE)
+    # Apply timetk index if selected
+    if (timetk_idx) {
+        idx <- tk_index(x, timetk_idx = TRUE)
         if (nrow(ret) != length(idx)) ret <- ret[(nrow(ret) - length(idx) + 1):nrow(ret),]
         ret[, rename_index] <- idx
     }
 
     # Index using sw_augment_columns() with data = NULL
-    ret <- sw_augment_columns(ret, data = NULL, rename_index = rename_index, timekit_idx = timekit_idx)
+    ret <- sw_augment_columns(ret, data = NULL, rename_index = rename_index, timetk_idx = timetk_idx)
 
     return(ret)
 }