+++ R/tidiers_decomposed_ts.R
@@ -4,8 +4,8 @@
 #' @param x An object of class "decomposed.ts"
 #' @param rename_index Used with `sw_augment` and `sw_tidy_decomp`.
 #' A string representing the name of the index generated.
-#' @param timekit_idx Used with `sw_augment` and `sw_tidy_decomp`.
-#' When `TRUE`, uses a timekit index (irregular, typically date or datetime) if present.
+#' @param timetk_idx Used with `sw_augment` and `sw_tidy_decomp`.
+#' When `TRUE`, uses a timetk index (irregular, typically date or datetime) if present.
 #' @param ... Not used.
 #'
 #'
@@ -37,13 +37,13 @@ NULL
 #'   * `seasadj`: observed - season
 #'
 #' @export
-sw_tidy_decomp.decomposed.ts <- function(x, timekit_idx = FALSE, rename_index = "index", ...) {
+sw_tidy_decomp.decomposed.ts <- function(x, timetk_idx = FALSE, rename_index = "index", ...) {
 
-    # Check timekit_idx
-    if (timekit_idx) {
-        if (!has_timekit_idx(x$x)) {
-            warning("Object has no timekit index. Using default index.")
-            timekit_idx = FALSE
+    # Check timetk_idx
+    if (timetk_idx) {
+        if (!has_timetk_idx(x$x)) {
+            warning("Object has no timetk index. Using default index.")
+            timetk_idx = FALSE
         }
     }
 
@@ -56,15 +56,15 @@ sw_tidy_decomp.decomposed.ts <- function(x, timekit_idx = FALSE, rename_index = 
     # Coerce to tibble
     ret <- tk_tbl(ret, preserve_index = TRUE, rename_index, silent = TRUE)
 
-    # Apply timekit index if selected
-    if (timekit_idx) {
-        idx <- tk_index(x$x, timekit_idx = TRUE)
+    # Apply timetk index if selected
+    if (timetk_idx) {
+        idx <- tk_index(x$x, timetk_idx = TRUE)
         if (nrow(ret) != length(idx)) ret <- ret[(nrow(ret) - length(idx) + 1):nrow(ret),]
         ret[, rename_index] <- idx
     }
 
     # Index using sw_augment_columns() with data = NULL
-    ret <- sw_augment_columns(ret, data = NULL, rename_index = rename_index, timekit_idx = timekit_idx)
+    ret <- sw_augment_columns(ret, data = NULL, rename_index = rename_index, timetk_idx = timetk_idx)
 
     return(ret)
 }