+++ R/tidiers_stl.R
@@ -4,8 +4,8 @@
 #' @param x An object of class "stl" or "stlm"
 #' @param rename_index Used with `sw_tidy_decomp`.
 #' A string representing the name of the index generated.
-#' @param timekit_idx Used with `sw_tidy_decomp`.
-#' When `TRUE`, uses a timekit index (irregular, typically date or datetime) if present.
+#' @param timetk_idx Used with `sw_tidy_decomp`.
+#' When `TRUE`, uses a timetk index (irregular, typically date or datetime) if present.
 #' @param ... Not used.
 #'
 #'
@@ -47,13 +47,13 @@ sw_tidy.stl <- function(x, ...) {
 #'   * `seasadj`: observed - season (or trend + remainder)
 #'
 #' @export
-sw_tidy_decomp.stl <- function(x, timekit_idx = FALSE, rename_index = "index", ...) {
+sw_tidy_decomp.stl <- function(x, timetk_idx = FALSE, rename_index = "index", ...) {
 
-    # Check timekit_idx
-    if (timekit_idx) {
-        if (!has_timekit_idx(x)) {
-            warning("Object has no timekit index. Using default index.")
-            timekit_idx = FALSE
+    # Check timetk_idx
+    if (timetk_idx) {
+        if (!has_timetk_idx(x)) {
+            warning("Object has no timetk index. Using default index.")
+            timetk_idx = FALSE
         }
     }
 
@@ -67,15 +67,15 @@ sw_tidy_decomp.stl <- function(x, timekit_idx = FALSE, rename_index = "index", .
     # Coerce to tibble
     ret <- tk_tbl(ret, preserve_index = TRUE, rename_index, silent = TRUE)
 
-    # Apply timekit index if selected
-    if (timekit_idx) {
-        idx <- tk_index(x, timekit_idx = TRUE)
+    # Apply timetk index if selected
+    if (timetk_idx) {
+        idx <- tk_index(x, timetk_idx = TRUE)
         if (nrow(ret) != length(idx)) ret <- ret[(nrow(ret) - length(idx) + 1):nrow(ret),]
         ret[, rename_index] <- idx
     }
 
     # Index using sw_augment_columns() with data = NULL
-    ret <- sw_augment_columns(ret, data = NULL, rename_index = rename_index, timekit_idx = timekit_idx)
+    ret <- sw_augment_columns(ret, data = NULL, rename_index = rename_index, timetk_idx = timetk_idx)
 
     return(ret)
 }
@@ -84,9 +84,9 @@ sw_tidy_decomp.stl <- function(x, timekit_idx = FALSE, rename_index = "index", .
 #' @rdname tidiers_stl
 #'
 #' @export
-sw_tidy_decomp.stlm <- function(x, timekit_idx = FALSE, rename_index = "index", ...) {
+sw_tidy_decomp.stlm <- function(x, timetk_idx = FALSE, rename_index = "index", ...) {
 
-    ret <- sw_tidy_decomp(x$stl, timekit_idx = timekit_idx, rename_index = rename_index, ...)
+    ret <- sw_tidy_decomp(x$stl, timetk_idx = timetk_idx, rename_index = rename_index, ...)
 
     return(ret)
 }