+++ tests/testthat/test_tidiers_lm.R
@@ -1,7 +1,7 @@
 library(sweep)
 library(forecast)
 library(tidyquant)
-library(timekit)
+library(timetk)
 context("Testing lm tidiers")